Crate rustorio

The first game written and played entirely in Rust’s type system. Not just do you play by writing Rust code, the rules of the game are enforced by the Rust compiler! If you can write the program so it compiles and doesn’t panic, you win!

How to play

Install Rust. Specifically it’s important to have the entire rustup toolchain and cargo, all of which you get automatically by following the instructions in the link.

Install rustorio by running cargo +nightly install rustorio.

Set up a new Rustorio project by running rustorio setup <path>, where <path> is the directory you want to create the project in (defaults to ‘.’).

Under src/bin/tutorial/ you will find a tutorial save. You can start by playing that one.

Playing the game consists of filling out the user_main function in the main.rs file in the save game folder created for you.

Run with rustorio play <save name>. This will compile and run your save. If it compiles and completes without panicking, you win! It’ll then tell you how many ticks it took you to win.

After the tutorial

To play other game modes, run rustorio new-game and specify a game mode. Use rustorio new-game --help to see all available game modes.

Rules

The rules are enforced by the compiler. Only thing I’d say is to not remove the #![forbid(unsafe_code)] at the top of the main.rs file, as using unsafe code can bypass most of what the compiler enforces. If you think you’ve found a way to cheat the game that isn’t caught by the compiler, please file an issue! Part of my interest in this project is whether I can actually rule out all possible cheating vectors using only the Rust compiler. I haven’t found a way to cheat yet, but I’m sure others will be more creative than me!

Help

Documentation for the Rustorio library can be found here. A good place to start is to build a furnace and start mining and smelting iron. Alternatively, you can work backwards by looking at the recipe for points to figure out how to get them.

Modules

buildings

Buildings take inputs to produce outputs over time.

gamemodes

recipes

A recipe is a way of turning resources into other resources. A specific recipe specifies the input and output resources, as well as the time it takes to complete the recipe.

research

Structs

Bundle

Contains a fixed (compile-time known) amount of a resource. A Bundle can be used to build structures or as input for recipes.

Resource

Holds an arbitrary amount of a resource. A Resource object can be split into smaller parts, combined or Bundles can be extracted from them.

Tick

The tick is used to keep track of time in the game. You can advance the game by one tick using the next method. Many functions and building methods require a Tick to be passed in, which allows them to update their state. If a function takes a &mut Tick, then the function will take time. If a function merely takes a &Tick, or no Tick at all, it will never advance the game time.

Enums

ResourceType

The different types of resources in the game. Used as a const generic parameter for both Resource and Bundle.

Functions

mine_copper

Mines copper ore. Takes 2 ticks to mine 1 ore.

mine_iron

Mines iron ore. Takes 2 ticks to mine 1 ore.

play

Runs your play. If it is run multiple times, it will panic. This is to prevent using multiple threads to cheat.

Type Aliases

CopperOreBundle

A bundle of specifically copper ore.

IronOreBundle

A bundle of specifically iron ore.