rustorio::buildings

Struct Furnace

pub struct Furnace<R: FurnaceRecipe> { /* private fields */ }

The furnace is used to smelt ores into base resources.

Implementations

impl<R: FurnaceRecipe> Furnace<R>

pub fn build(tick: &Tick, _iron: Bundle<{ ResourceType::Iron }, 10>) -> Self

Builds a furnace. Costs 10 iron.

pub fn change_recipe<R2: FurnaceRecipe>(self) -> Result<Furnace<R2>, Furnace<R>>

Changes the Recipe of the furnace. Returns the original furnace if the furnace has no inputs or outputs.

pub fn cur_input(&mut self, tick: &Tick) -> u32

How much of the input resource is currently in the furnace.

pub fn cur_output(&mut self, tick: &Tick) -> u32

How much of the output resource is currently in the furnace.

pub fn add_input<const AMOUNT: u32>( &mut self, tick: &Tick, _ore: Bundle<{ R::INPUT }, AMOUNT>, )

Consumes a Bundle and puts the contained resources into the furnace.

pub fn take_input<const AMOUNT: u32>( &mut self, tick: &Tick, ) -> Option<Bundle<{ R::INPUT }, AMOUNT>>

Takes a specified amount of input resources from the furnace and puts it into a Bundle.

pub fn empty_input(&mut self, tick: &Tick) -> Resource<{ R::INPUT }>

Takes all input resources currently in the furnace and puts it into a Resource.

pub fn take_output<const AMOUNT: u32>( &mut self, tick: &Tick, ) -> Option<Bundle<{ R::OUTPUT }, AMOUNT>>

Takes a specified amount of output resources from the furnace and puts it into a Bundle.

pub fn empty_output(&mut self, tick: &Tick) -> Resource<{ R::OUTPUT }>

Takes all output resources currently in the furnace and puts it into a Resource.