rustorio

Module gamemodes

Structs

Standard

The standard game mode. Forces the player through the main gameplay mechanics.

StandardStartingResources

Starting resources for the standard game mode. Includes 10 iron and the ability to research points.

Tutorial

The tutorial game mode. Very short distance from start to victory, meant to teach the very most basic elements of play.

TutorialStartingResources

Starting resources for the tutorial game mode.

Traits

GameMode

A game mode defines the starting resources and victory conditions for a game.