rustorio::gamemodes

Trait GameMode

pub trait GameMode {
    type StartingResources: StartingResources;
    type VictoryResources;
}

A game mode defines the starting resources and victory conditions for a game.

Required Associated Types

type StartingResources: StartingResources

Starting resources provided to the player at the beginning of the game.

type VictoryResources

Resources required to achieve victory.