rustorio::recipes

Trait AssemblerRecipe

pub trait AssemblerRecipe: Debug + Sealed {
    const INPUT1: ResourceType;
    const INPUT1_AMOUNT: u32;
    const INPUT2: ResourceType;
    const INPUT2_AMOUNT: u32;
    const OUTPUT: ResourceType;
    const OUTPUT_AMOUNT: u32;
    const TIME: u64;
}

Any recipe that implements this trait can be used in an Assembler.

Required Associated Constants

const INPUT1: ResourceType

The first of the two types of resources consumed by this recipe.

const INPUT1_AMOUNT: u32

The amount of the first input resource consumed by this recipe.

const INPUT2: ResourceType

The second of the two types of resources consumed by this recipe.

const INPUT2_AMOUNT: u32

The amount of the second input resource consumed by this recipe.

const OUTPUT: ResourceType

The type of resource produced by this recipe.

const OUTPUT_AMOUNT: u32

The amount of the output resource produced by this recipe.

const TIME: u64

The time (in ticks) it takes to complete this recipe.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.