rustorio::research

Trait Research

pub trait Research: Sealed {
    type Unlocks;

    const RED_SCIENCE_COST: u32;

    // Required method
    fn research(
        self,
        _red_science: Bundle<{ ResourceType::RedScience }, AMOUNT>,
    ) -> Self::Unlocks;
}

A research can be carried out by calling the research method with the required science packs. This will consume the science packs and the research itself, and return whatever the research unlocks, mostly recipes and other researches.

Required Associated Constants

const RED_SCIENCE_COST: u32

The amount of red science required to carry out this research.

Required Associated Types

type Unlocks

The reward for completing this research.

Required Methods

fn research( self, _red_science: Bundle<{ ResourceType::RedScience }, AMOUNT>, ) -> Self::Unlocks

Carries out the research by consuming the required science packs and the research itself, returning whatever this research unlocks.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.